Renamed TweeterController to FollowController. Used POST and DELETE methods for FollowUser and UnfollowUser calls, respectively.
