+++ Tweeter/Tweeter.csproj
@@ -181,7 +181,7 @@
     <Compile Include="Controllers\HomeController.cs" />
     <Compile Include="Controllers\ManageController.cs" />
     <Compile Include="Controllers\TweetController.cs" />
-    <Compile Include="Controllers\TweeterController.cs" />
+    <Compile Include="Controllers\FollowController.cs" />
     <Compile Include="Controllers\TwitUsernameController.cs" />
     <Compile Include="DAL\TweeterContext.cs" />
     <Compile Include="DAL\TweeterRepository.cs" />
